fix(certificate): repair number allocation and first-time settings save
