+++ admin_core_service/src/main/java/vacademy/io/admin_core_service/features/institute/service/setting/CertificateSettingStrategy.java
@@ -2,6 +2,7 @@ package vacademy.io.admin_core_service.features.institute.service.setting;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import vacademy.io.admin_core_service.features.institute.constants.ConstantsSettingDefaultValue;
 import vacademy.io.admin_core_service.features.institute.dto.settings.InstituteSettingDto;
 import vacademy.io.admin_core_service.features.institute.dto.settings.SettingDto;
